+++ src/roi_queue.rs
@@ -1,5 +1,5 @@
 use crossbeam_queue::SegQueue;
-use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 
 /// A reuse-or-increment queue for unique identifiers. When polled for a new identifier, this will fist attempt to reuse any
 /// relinquished identifiers, before incrementing to create a new identifier if no other could be found. This is lock-free, and
@@ -10,15 +10,15 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 pub(crate) struct RoiQueue {
     /// A counter that represents one above the highest active identifier. In other words, the value of this counter will always
     /// be an identifier that is not yet in circulation.
-    counter: AtomicUsize,
+    counter: AtomicU32,
     /// A lock-free queue containing all the identifiers that have been 'relinquished' (i.e. that are no longer in use). These
     /// will be recirculated automatically when a new identifier is requested.
-    relinquished: SegQueue<usize>,
+    relinquished: SegQueue<u32>,
 }
 impl Default for RoiQueue {
     fn default() -> Self {
         Self {
-            counter: AtomicUsize::new(0),
+            counter: AtomicU32::new(0),
             relinquished: SegQueue::new(),
         }
     }
@@ -30,7 +30,7 @@ impl RoiQueue {
     }
     /// Gets a new unique identifier from this queue. This method guarantees that the returned identifier will be unique among
     /// those returned from *this* queue.
-    pub fn get(&self) -> usize {
+    pub fn get(&self) -> u32 {
         // Attempt to acquire a relinquished ID from the queue
         if let Some(id) = self.relinquished.pop() {
             id
@@ -41,7 +41,7 @@ impl RoiQueue {
     }
     /// Marks the given identifier as relinquished. After this method is called, the given identifier will almost certainly
     /// point to something else, and as such it should no longer be considered valid.
-    pub fn relinquish(&self, id: usize) {
+    pub fn relinquish(&self, id: u32) {
         self.relinquished.push(id);
     }
 }
